Started adding UI for the Client
The client will have a simple UI with a message thread and a way of writing the message. Also, some way of setting the username should be created.

The `client.sh` and `server.sh` are now renamed to simply `client` and `server` since *nix (Linux, Unix, etc.) allow everything to be an executable as long as you execute a `chmod +x` on it

TODO:
- Wire the UI with the Client Socket logic